// frontend/src/interfaces/userData.ts
export interface UserData {
    [key: string]: number | string
    name: string
    caloriesLimit: number
    carbsLimit: number
    proteinLimit: number
    fatsLimit: number
    breakfastCaloriesLimit: number
    lunchCaloriesLimit: number
    snacksCaloriesLimit: number
    dinnerCaloriesLimit: number
}